@@ -3,19 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "firebase": "^6.4.0",
-    "moment": "^2.24.0",
-    "react": "^16.9.0",
+    "firebase": "^10.5.2",
+    "moment": "^2.29.4",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
     "react-delay": "^0.1.0",
-    "react-dom": "^16.9.0",
-    "react-router-dom": "^5.0.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0",
     "react-router-last-location": "^2.0.1",
-    "react-scripts": "^3.4.1",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.87.3",
+    "react-script": "^2.0.5",
+    "react-scripts": "^5.0.1",
+    "semantic-ui-css": "^2.5.0",
+    "semantic-ui-react": "^2.1.4",
     "shortid": "^2.2.14",
     "store": "^2.0.12",
-    "uuid": "^3.3.2"
+    "uuid": "^9.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -39,7 +41,7 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.2"
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0"
   }
 }
