@@ -1,152 +1,118 @@
 // Theirs
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import React, {useEffect, useState} from 'react';
+import {Link} from 'react-router-dom';
 import moment from 'moment';
-import {
-  Button,
-  Container,
-  Divider,
-  Form,
-  Header,
-  List,
-  Segment,
-  Icon,
-} from 'semantic-ui-react';
+import {Container, Header, Icon, List, Segment,} from 'semantic-ui-react';
+import {onValue, set} from 'firebase/database';
+import shortid from 'shortid';
 
 // Ours
-import { db, auth } from '../../firebase';
+import {auth, db} from '../../firebase';
 import * as pokerTablesApi from '../../api/pokerTables';
 import Layout from '../../containers/Layout';
+import withAuthentication from '../../containers/withAuthentication';
+import PokerTableNameForm from './PokerTableNameForm';
 
-const shortid = require('shortid');
 
-class Dashboard extends Component {
-  state = {
-    pokerTables: [],
-    pokerTablesClient: null,
-    newPokerTableName: '',
-    currentUser: auth.getAuth().currentUser,
-  };
+const Dashboard = () => {
+	const currentUser = auth.auth.currentUser;
+	const pokerTablesClient = pokerTablesApi.createClient(currentUser.uid);
+	const [pokerTables, setPokerTables] = useState([]);
 
-  componentDidMount() {
-    this.loadPokerTables();
+	useEffect(() => {
+		loadPokerTables();
+	}, []);
 
-    this.setState({
-      pokerTablesClient: pokerTablesApi.createClient(
-        this.state.currentUser.uid
-      ),
-    });
-  }
+	const createPokerTable = (newPokerTableName) => {
+		const uid = shortid.generate();
+		const pRef = db.pokerTable(currentUser.uid, uid);
+		const data = {
+			created: new Date().toString(),
+			tableName: newPokerTableName,
+		};
+		set(pRef, data)
+			.then(() => console.log('Updated successfully'))
+			.catch((error) => console.log('Error updating document: ', error));
+		loadPokerTables();
+	};
 
-  createPokerTable = (e) => {
-    const pRef = db.pokerTablesRoot(this.state.currentUser.uid);
-    pRef.child(shortid.generate()).update({
-      tableName: this.state.newPokerTableName,
-      created: new Date(),
-    });
-    this.setState({ newPokerTableName: '' });
-    this.loadPokerTables();
-  };
+	const removePokerTable = (pokerTableId) => (e) => {
+		e.preventDefault();
 
-  removePokerTable = (pokerTableId) => (e) => {
-    e.preventDefault();
+		// Optimistically deletes poker table. i.e. doesn't block the ui from updating
+		pokerTablesClient.remove(pokerTableId);
 
-    this.state.pokerTablesClient.remove(pokerTableId); // Optimistically deletes poker table. i.e. doesn't block the ui from updating
+		const filteredPokerTables = pokerTables.filter(
+			({id}) => id !== pokerTableId
+		);
 
-    const filteredPokerTables = this.state.pokerTables.filter(
-      ({ id }) => id !== pokerTableId
-    );
+		setPokerTables(filteredPokerTables);
+	};
 
-    this.setState({
-      pokerTables: filteredPokerTables,
-    });
-  };
+	const loadPokerTables = () => {
+		const pokerTablesRef = db.pokerTables(currentUser.uid);
+		onValue(pokerTablesRef, (snapshot) => {
+			if (snapshot.exists()) {
+				const pokerTables = snapshot.val();
+				let newPokerTablesState = [];
+				for (let table in pokerTables) {
+					newPokerTablesState.push({
+						...pokerTables[table],
+						id: table,
+					});
+				}
+				newPokerTablesState.sort((t1, t2) => {
+					if (t1.created > t2.created) return -1;
+					if (t2.created > t1.created) return 1;
+					return 0;
+				});
+				setPokerTables(newPokerTablesState);
+			}
+		});
+	};
 
-  handleNewPokerTableName = (e) => {
-    this.setState({ newPokerTableName: e.target.value });
-  };
+	return (
+		<Layout data-testid={`Dashboard`}>
+			<Container>
+				<Segment raised>
+					<PokerTableNameForm handlePokerTableSubmit={createPokerTable}/>
+				</Segment>
+				<Segment stacked>
+					<Header as="h1">Your Poker Tables</Header>
+					<List divided relaxed>
+						{pokerTables.length > 0 ? pokerTables.map((s) => (
+							<List.Item key={s.id} className="pwm-list-item">
+								<Link
+									to={`/table/${currentUser.uid}/${s.id}`}
+									className="pwm-list-item-content"
+								>
+									<List.Content>
+										<List.Header>{s.tableName}</List.Header>
+										<List.Description>Table ID: {s.id}</List.Description>
+										<List.Description>
+											Created: {moment(s.created).format('MM/DD/YYYY hh:mma')}
+										</List.Description>
+									</List.Content>
+								</Link>
+								<div className="actions">
+									<button
+										className="pwm-delete"
+										data-testid="delete-button"
+										onClick={removePokerTable(s.id)}
+									>
+										<Icon name="times" color="red"/>
+									</button>
+								</div>
+							</List.Item>
+						)) : `No poker tables found. Create one above!`}
+					</List>
+				</Segment>
+			</Container>
+		</Layout>
+	);
+};
 
-  loadPokerTables = () => {
-    const pokerTablesRef = db.pokerTables(this.state.currentUser.uid);
-    pokerTablesRef.on('value', (snapshot) => {
-      let pokerTables = snapshot.val();
-      let newPokerTablesState = [];
-      for (let table in pokerTables) {
-        newPokerTablesState.push({
-          ...pokerTables[table],
-          id: table,
-        });
-      }
-      newPokerTablesState.sort((t1, t2) => {
-        if (t1.created > t2.created) return -1;
-        if (t2.created > t1.created) return 1;
-        return 0;
-      });
-      this.setState({
-        pokerTables: newPokerTablesState,
-      });
-    });
-  };
+export default withAuthentication(Dashboard);
 
-  render() {
-    return (
-      <Layout>
-        <Container>
-          <Segment raised>
-            <Form onSubmit={this.createPokerTable}>
-              <Header as="h1">Create Poker Table</Header>
-              <Form.Field>
-                <label>Poker Table Name</label>
-                <input
-                  placeholder="New Poker Table Name"
-                  value={this.state.newPokerTableName}
-                  onChange={this.handleNewPokerTableName}
-                />
-              </Form.Field>
-              <Button primary type="submit">
-                Create Poker Table
-              </Button>
-            </Form>
-          </Segment>
-          <Segment stacked>
-            <Header as="h1">Your Poker Tables</Header>
-            <List divided relaxed>
-              {this.state.pokerTables.map((s) => (
-                <List.Item key={s.id} className="pwm-list-item">
-                  <Link
-                    to={`/table/${this.state.currentUser.uid}/${s.id}`}
-                    className="pwm-list-item-content"
-                  >
-                    <List.Content>
-                      <List.Header>{s.tableName}</List.Header>
-                      <List.Description>Table ID: {s.id}</List.Description>
-                      <List.Description>
-                        Created: {moment(s.created).format('MM/DD/YYYY hh:mma')}
-                      </List.Description>
-                    </List.Content>
-                  </Link>
-                  <div className="actions">
-                    <button
-                      className="pwm-delete"
-                      onClick={this.removePokerTable(s.id)}
-                    >
-                      <Icon name="times" color="red" />
-                    </button>
-                  </div>
-                </List.Item>
-              ))}
-            </List>
-          </Segment>
-        </Container>
-        <Divider horizontal></Divider>
-        <Container>
-          <Button negative onClick={() => auth.getAuth().signOut()}>
-            Logout
-          </Button>
-        </Container>
-      </Layout>
-    );
-  }
-}
-
-export default Dashboard;
+// Export Dashboard without the withAuthentication HOC for testing
+export {Dashboard};
