@@ -1,79 +1,103 @@
-// Theirs
 import React from 'react';
-import { MemoryRouter } from 'react-router-dom';
-import { render, fireEvent } from '@testing-library/react';
-
-// Ours
+import {fireEvent, render, waitFor} from '@testing-library/react';
+import '@testing-library/jest-dom';
+import {onValue, set} from 'firebase/database';
+import {Dashboard} from './index';
 import * as pokerTablesApi from '../../api/pokerTables';
-import Dashboard from '.';
+import {MemoryRouter} from 'react-router-dom';
 
-jest.mock('../../firebase', () => {
-  return {
-    auth: {
-      getAuth: jest.fn().mockReturnValue({
-        currentUser: {
-          uid: 'userId',
-        },
-      }),
-    },
-    db: {
-      pokerTables: jest.fn().mockReturnValue({
-        on: jest.fn().mockImplementation((_, cb) => {
-          const snapshot = {
-            val: jest.fn().mockReturnValue({
-              pt1: { tableName: 'table 1', created: new Date() },
-              pt2: { tableName: 'table 2', created: new Date() },
-              pt3: { tableName: 'table 3', created: new Date() },
-            }),
-          };
+// Mock necessary modules and functions
+// Mock the firebase functions
+jest.mock('firebase/database');
+jest.mock('../../firebase');
+jest.mock('../../api/pokerTables');
 
-          cb(snapshot);
-        }),
-      }),
-    },
-  };
+// mock layout component
+jest.mock('../../containers/Layout', () => ({
+	__esModule: true,
+	default: ({children}) => <div data-testid="layout-test">{children}</div>,
+}));
+
+jest.mock('../../firebase', () => ({
+	...jest.requireActual('../../firebase'),
+	auth: {
+		get auth() {
+			return {
+				currentUser: {uid: 'testUserId'},
+			};
+		},
+	},
+	db: {
+		pokerTables: jest.fn((userId) => ({
+			path: `pokerTables/${userId}`,
+		})),
+		pokerTable: jest.fn((userId, tableId) => ({
+			path: `pokerTable/${userId}/${tableId}`,
+		})),
+	},
+}));
+jest.mock(pokerTablesApi, () => ({
+	createClient: jest.fn(() => ({
+		remove: jest.fn(),
+	})),
+}));
+jest.mock('moment', () => {
+	const moment = jest.requireActual('moment');
+	return (date) => moment(date);
 });
-jest.mock('../../api/pokerTables', () => {
-  return {
-    createClient: jest.fn().mockReturnValue({
-      remove: jest.fn().mockResolvedValue({}),
-    }),
-  };
-});
+jest.mock('shortid', () => ({generate: jest.fn(() => 'testTableId')}));
 
-describe('dashboard page', () => {
-  it('can delete a poker table', () => {
-    expect.hasAssertions();
+describe('Dashboard Page', () => {
+	test('renders Dashboard component', async () => {
+		// Render the component
+		const {getByText} = render(<Dashboard/>);
 
-    const { baseElement: el } = render(
-      <MemoryRouter>
-        <Dashboard />
-      </MemoryRouter>
-    );
+		// Verify that the component renders correctly
+		expect(getByText('Your Poker Tables')).toBeInTheDocument();
+	});
 
-    let listItems = el.querySelectorAll('.pwm-list-item');
+	test('deletes a poker table when the delete button is clicked', async () => {
+		// Mock necessary functions
+		const remove = jest.fn();
+		onValue.mockImplementationOnce((ref, callback) => {
+			const snapshot = {
+				exists: jest.fn(() => true),
+				val: jest.fn(() => ({
+					'randomuserstring': {
+						tableName: 'Test Table',
+						created: 'Fri Nov 17 2023 22:31:08 GMT-0500 (Eastern Standard Time)'
+					}
+				}))
+			};
+			callback(snapshot);
+		});
+		set.mockImplementationOnce(() => Promise.resolve());
+		pokerTablesApi.createClient.mockImplementation(() => ({ remove }));
 
-    // sanity check
-    expect(listItems).toHaveLength(3);
-    expect(el).toHaveTextContent('table 2');
+		// Render the component inside memory router
+		const {getByTestId, getByText, queryByText} = await render(
+			<MemoryRouter>
+				<Dashboard/>
+			</MemoryRouter>
+		);
 
-    const deleteButton = listItems[1].querySelector('.pwm-delete');
+		// Wait for data to be loaded
+		await waitFor(() => {
+			expect(getByText('Test Table')).toBeInTheDocument();
+		});
 
-    fireEvent.click(deleteButton);
+		// Click the delete button
+		fireEvent.click(getByTestId('delete-button'));
 
-    // ensure the api call was made
-    const mockedPokerTablesClient =
-      pokerTablesApi.createClient.mock.results[0].value;
+		// Verify that the poker table is deleted optimistically
+		expect(remove).toHaveBeenCalledWith('randomuserstring');
 
-    expect(mockedPokerTablesClient.remove).toHaveBeenCalledWith('pt2');
+		// Verify that the poker table is removed from the UI
+		await waitFor(() => {
+			const table = queryByText('Test Table');
+			expect(table).not.toBeInTheDocument();
+		});
+	});
 
-    listItems = el.querySelectorAll('.pwm-list-item'); // reload selection
-
-    expect(listItems).toHaveLength(2);
-    expect(el).not.toHaveTextContent('table 2');
-
-    // ensure we have the correct order of items
-    expect(listItems[0]).toHaveTextContent('table 1');
-    expect(listItems[1]).toHaveTextContent('table 3');
-  });
+	// Add more tests based on your specific requirements
 });
