@@ -1,232 +1,176 @@
-import React, {Component} from 'react';
-import {
-  Button,
-  Card,
-  Container,
-  Divider,
-  Header,
-  Icon,
-  Loader,
-  Segment,
-} from 'semantic-ui-react';
-import { db, auth } from '../../firebase';
+import React, {useEffect, useState} from 'react';
+import {Card, Container, Header, Loader, Segment,} from 'semantic-ui-react';
+import {auth, db} from '../../firebase';
 import './issue.css';
+import {useParams} from 'react-router-dom';
+import {child, onValue, update} from 'firebase/database';
+import VotingBlock from './VotingBlock';
+import Controls from './Controls';
 
-const availablePoints = [0,1,2,3,5,8,13,21];
+const Issue = ({issue}) => {
+	const {userId, tableId} = useParams();
+	const currentUser = auth.auth.currentUser;
+	const isTableOwner = userId === currentUser.uid;
+	const [issueState, setIssueState] = useState({
+		isLoaded: false,
+		isLocked: false,
+		showVotes: false,
+		title: '',
+	});
+	const [votesState, setVotesState] = useState({
+		isLoaded: false,
+		mostVotes: -1,
+		userVote: null,
+		votes: [],
+	});
+	const issueRef = db.pokerTableIssue(
+		userId,
+		tableId,
+		issue
+	);
+	const votesRef = db.votesRoot(
+		issue
+	);
 
-class Issue extends Component {
-  state = {
-    ownerId: this.props.ownerId,
-    tableId: this.props.tableId,
-    currentUser: auth.getAuth().currentUser,
-    currentIssueId: this.props.issue,
-    title: '',
-    votes: [],
-    isLocked: false,
-    showVotes: false,
-    userVote: null,
-    mostVotes: null,
-    isLoaded: false,
-    isTableOwner: false,
-  };
+	useEffect(() => {
+		return loadIssue();
+	}, []);
 
-  componentDidMount() {
-    this.issueRef = db.pokerTableIssue(
-      this.state.ownerId,
-      this.state.tableId,
-      this.state.currentIssueId
-    );
-    this.votesRef = db.votesRoot(
-      this.state.currentIssueId
-    );
-    this.loadIssue();
-    this.loadVotes();
-  }
+	useEffect(() => {
+		return loadVotes();
+	}, []);
 
-  loadIssue() {
-    this.issueRef.on('value', snapshot => {
-      const issue = snapshot.val();
+	const loadIssue = () => onValue(issueRef, snapshot => {
+		if (snapshot.exists()) {
+			const issue = snapshot.val();
+			const newState = {
+				...issueState,
+				title: issue.title,
+				isLocked: issue.isLocked || false,
+				showVotes: issue.showVotes || false,
+				isLoaded: true,
+			};
+			setIssueState(newState);
+		}
+	});
 
-      this.setState({
-        title: issue.title,
-        isLocked: issue.isLocked || false,
-        showVotes: issue.showVotes || false,
-        isLoaded: true,
-        isTableOwner: this.state.ownerId === this.state.currentUser.uid,
-      });
-    });
-  }
+	const loadVotes = () => onValue(votesRef, snapshot => {
+		if (snapshot.exists()) {
+			const newVotesList = [];
+			const votes = snapshot.val() || {};
+			for (let vote in votes) {
+				newVotesList.push({
+					...votes[vote],
+					userId: vote,
+				});
+			}
+			newVotesList.sort((v1, v2) => {
+				if (v1.vote > v2.vote) return 1;
+				if (v2.vote > v1.vote) return -1;
+				return 0;
+			});
 
-  loadVotes() {
-    this.votesRef.on('value', snapshot => {
-      const newVotesList = [];
-      const votes = snapshot.val() || {};
-      for (let vote in votes) {
-        newVotesList.push({
-          ...votes[vote],
-          userId: vote,
-        });
-      }
-      newVotesList.sort( (v1, v2) => {
-        if(v1.vote > v2.vote) return 1;
-        if(v2.vote > v1.vote) return -1;
-        return 0;
-      });
+			// Get most votes
+			const voteTally = newVotesList.reduce((acc, curr) => {
+				if (curr.vote in acc) {
+					acc[curr.vote]++;
+				} else {
+					acc[curr.vote] = 1;
+				}
+				return acc;
+			}, {});
 
-      // Get most votes
-      const voteTally = newVotesList.reduce((acc, curr) => {
-        if (curr.vote in acc) {
-            acc[curr.vote]++;
-        } else {
-            acc[curr.vote] = 1;
-        }
-        return acc;
-      }, {});
+			let mostVotes = -1;
+			let multipleModes = false;
+			for (let points in voteTally) {
+				let currentMostVotes = voteTally[mostVotes] || 0;
+				if (voteTally[points] === currentMostVotes) {
+					multipleModes = true;
+				} else if (voteTally[points] >= currentMostVotes) {
+					mostVotes = parseInt(points, 10);
+					multipleModes = false;
+				}
+			}
+			if (multipleModes) {
+				// don't highlight any point values
+				mostVotes = -1;
+			}
 
-      let mostVotes = -1;
-      let multipleModes = false;
-      for(let points in voteTally) {
-        let currentMostVotes = voteTally[mostVotes] || 0;
-        if(voteTally[points] === currentMostVotes) {
-          multipleModes = true;
-        } else if(voteTally[points] >= currentMostVotes) {
-          mostVotes = parseInt(points, 10);
-          multipleModes = false;
-        }
-      }
-      if(multipleModes) {
-        // don't highlight any point values
-        mostVotes = -1;
-      }
+			const myVote =
+				newVotesList.find(v => v.userId === currentUser.uid);
+			const newState = {
+				...votesState,
+				userVote: (myVote) ? myVote.vote : null,
+				votes: newVotesList.length ? newVotesList : [],
+				mostVotes
+			};
+			// setState(newState);
+			setVotesState(newState);
+		} else {
+			console.warn('No votes found');
+		}
+	});
 
-      const myVote =
-        newVotesList.find( v => v.userId === this.state.currentUser.uid);
-      this.setState({
-        userVote: (myVote) ? myVote.vote : null,
-        votes: newVotesList,
-        mostVotes
-      });
-    });
-  }
+	const handleSelectVote = (userVote) => {
+		if (userVote === votesState.userVote) {
+			userVote = null;
+		}
+		setVotesState({
+			...votesState,
+			userVote
+		});
+		update(child(votesRef, currentUser.uid), {vote: userVote});
+	};
 
-  handleSelectVote(userVote) {
-    if(userVote === this.state.userVote) {
-      userVote = null;
-    }
-    this.setState({userVote});
-    this.votesRef.child(this.state.currentUser.uid)
-      .update({vote: userVote});
-  }
+	//suggestion() {
+	//let suggestion = '??';
+	//let mode = '??';
+	//if(issueState.showVotes) {
+	//const total = votesState.votes.reduce((t, v) => t + v.vote, 0);
+	//const suggestionAvg = (total / votesState.votes.length);
+	//suggestion = availablePoints.find( p => p >= suggestionAvg);
+	//mode = (votesState.mostVotes > -1) ? votesState.mostVotes : '--';
+	//}
+	//return(
+	//<Header sub>Mode/Mean ({mode}/{suggestion})</Header>
+	//);
+	//}
 
-  handleShow() {
-    this.issueRef.child('showVotes').set(!this.state.showVotes);
-  }
+	if (!issueState.isLoaded) {
+		return (<Loader size="large">Loading</Loader>);
+	}
 
-  handleLock() {
-    this.issueRef.child('isLocked').set(!this.state.isLocked);
-  }
-
-  votingBlock() {
-    if(this.state.isLocked) {
-      return;
-    }
-    return(
-      <Segment raised textAlign='center'>
-        <Card.Group
-          itemsPerRow={2}
-          id="voteCards"
-          className={(this.state.isLocked) ? 'locked' : 'unlocked'}
-        >
-          {availablePoints.map( p => (
-            <Card
-              key={p}
-              onClick={() => this.handleSelectVote(p)}
-              color='blue'
-              className={(this.state.userVote === p) ? 'selected' : ''}>
-              {p}
-            </Card>
-          ))}
-        </Card.Group>
-      </Segment>
-    );
-  }
-
-  //suggestion() {
-    //let suggestion = '??';
-    //let mode = '??';
-    //if(this.state.showVotes) {
-      //const total = this.state.votes.reduce((t, v) => t + v.vote, 0);
-      //const suggestionAvg = (total / this.state.votes.length);
-      //suggestion = availablePoints.find( p => p >= suggestionAvg);
-      //mode = (this.state.mostVotes > -1) ? this.state.mostVotes : '--';
-    //}
-    //return(
-      //<Header sub>Mode/Mean ({mode}/{suggestion})</Header>
-    //);
-  //}
-
-  controls() {
-    if(this.state.ownerId !== this.state.currentUser.uid) {
-      return;
-    }
-
-    return(
-      <Container id='voteControls' textAlign='center'>
-        <Button
-          positive
-          toggle
-          active={this.state.showVotes}
-          onClick={() => this.handleShow()}
-        >
-          <Icon
-            name={(this.state.showVotes) ? 'eye slash' : 'eye'}
-            size='large'/>
-            { ( () => (this.state.showVotes) ? 'Hide' : 'Show' )()} Votes
-        </Button>
-        <Button
-          negative
-          toggle
-          active={this.state.isLocked}
-          onClick={() => this.handleLock()}
-        >
-          <Icon
-            name={(this.state.isLocked) ? 'unlock' : 'lock'}
-            size='large' />
-            { ( () => (this.state.isLocked) ? 'Unlock' : 'Lock' )()} Voting
-        </Button>
-
-        <Divider horizontal />
-      </Container>
-    );
-  }
-
-  render() {
-    if(!this.state.isLoaded) {
-        return(<Loader size='large'>Loading</Loader>);
-    }
-
-    return (
-      <Container textAlign='center' id="issue">
-        <Header as='h1'>{this.state.title}</Header>
-        <Segment stacked>
-          {this.controls()}
-          <Card.Group
-            itemsPerRow={4}
-            id="voteCards"
-          >
-            {this.state.votes.map((v) => (
-              <Card color={(this.state.mostVotes === v.vote && this.state.showVotes) ? 'green' : 'blue'}
-                className={(this.state.mostVotes === v.vote && this.state.showVotes) ? 'mode' : ''}
-                key={v.userId}>
-                {(this.state.showVotes) ? v.vote : '?'}
-              </Card>
-            ))}
-          </Card.Group>
-        </Segment>
-        {this.votingBlock()}
-        </Container>
-    );
-  }
-}
+	return (
+		<Container textAlign="center" id="issue">
+			<Header as="h1">{issueState.title}</Header>
+			<Segment stacked>
+				{(isTableOwner) ?
+					<Controls
+						isLocked={issueState.isLocked}
+						issue={issue}
+						showVotes={issueState.showVotes}
+					/> : null}
+				<Card.Group
+					itemsPerRow={4}
+					id="voteCards"
+				>
+					{votesState.votes?.map((v) => (
+						<Card color={(votesState.mostVotes === v.vote && issueState.showVotes) ? 'green' : 'blue'}
+							  className={(votesState.mostVotes === v.vote && issueState.showVotes) ? 'mode' : ''}
+							  key={v.userId}>
+							{(issueState.showVotes) ? v.vote : '?'}
+						</Card>
+					))}
+				</Card.Group>
+			</Segment>
+			{!issueState.isLocked ?
+				<VotingBlock
+					isLocked={issueState.isLocked}
+					onClick={handleSelectVote}
+					userVote={votesState.userVote}
+				/> : null}
+		</Container>
+	);
+};
 
 export default Issue;
