@@ -1,5 +1,5 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import React from 'react';
+import {BrowserRouter, createBrowserRouter, Route, RouterProvider, Routes} from 'react-router-dom';
 
 import Login from '../components/Login';
 import Dashboard from '../components/Dashboard';
@@ -8,22 +8,15 @@ import About from '../components/About';
 import withAuthentication from '../containers/withAuthentication';
 import '../style.css';
 
-class App extends Component {
-  render() {
-    return (
-      <Router>
-        <Switch>
-          <Route path="/" exact component={Login} />
-          <Route path="/dashboard" component={withAuthentication(Dashboard)} />
-          <Route
-            path="/table/:userId/:tableId"
-            component={withAuthentication(PokerTable)}
-          />
-          <Route path="/about" component={About} />
-        </Switch>
-      </Router>
-    );
-  }
+const router = createBrowserRouter([
+    {path: "/", Component: Login},
+    {path: "/about", Component: About},
+    {path: "/dashboard", Component: Dashboard},
+    {path: "/table/:userId/:tableId", Component: PokerTable},
+])
+
+const App = () => {
+    return <RouterProvider router={router} />;
 }
 
 export default App;
